fix possible bugs

+++ src/HisabDo.Application/Services/CustomerService.cs
@@ -33,10 +33,15 @@ public class CustomerService(ICustomerRepository repository) : ICustomerService
         return ToDto(customer);
     }
 
-    public async Task<CustomerDto> UpdateAsync(int id, CreateCustomerDto dto)
+    public async Task<CustomerDto> UpdateAsync(int userId, int id, CreateCustomerDto dto)
     {
         var customer = await repository.GetByIdAsync(id)
             ?? throw new KeyNotFoundException($"No customer found with ID: {id}");
+
+        if (customer.UserId != userId)
+        {
+            throw new UnauthorizedAccessException("You do not have access to this customer.");
+        }
 
         customer.Name = dto.Name;
         customer.Phone = dto.Phone;
@@ -47,10 +52,15 @@ public class CustomerService(ICustomerRepository repository) : ICustomerService
         return ToDto(customer);
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task DeleteAsync(int userId, int id)
     {
         var customer = await repository.GetByIdAsync(id)
             ?? throw new KeyNotFoundException($"No customer found with ID: {id}");
+
+        if (customer.UserId != userId)
+        {
+            throw new UnauthorizedAccessException("You do not have access to this customer.");
+        }
 
         await repository.RemoveAsync(customer);
     }
